pyproject.toml
[project]
name = "techcamp-team-vision"
version = "1.0.0"
description = "Live Face Anti-Spoofing Demo — Techcamp 2026 Team Vision"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "pillow>=10.0.0",
    "timm>=0.9.0",
    "tqdm>=4.65.0",
    "scikit-learn>=1.3.0",
    "albumentations>=1.3.0",
]

[project.optional-dependencies]
gpu = [
    "insightface>=0.7.3",
    "onnxruntime-gpu==1.18.0",
]
mps = [
    "insightface>=0.7.3",
    "onnxruntime>=1.18.0",
]
gdrive = [
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.26.0",
    "pre-commit>=3.6.0",
    "ruff>=0.7.3",
    "black>=24.10.0",
    "isort>=5.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "nets"]

[tool.ruff]
exclude = ["nets/"]
line-length = 88

[tool.black]
line-length = 88
exclude = "nets/"

[tool.isort]
profile = "black"
skip = ["nets/"]
